app/requirements.txt
boto3
python-dotenv
uvicorn
fastapi
Pillow
